+++ examples/encryptionExample/src/main/java/io/realm/examples/encryption/MyApplication.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.realm.examples.encryptionexample;
+package io.realm.examples.encryption;
 
 import android.app.Application;